moved renderer into view
